Updates to the attack classes
