+++ Godot-Project/Assets/Scripts/Attacks/RepetativeAttackDirector.cs
@@ -1,5 +1,6 @@
 using Godot;
 using System;
+using System.Collections.Generic;
 
 public partial class RepetativeAttackDirector : AttackDirector
 {
@@ -8,6 +9,8 @@ public partial class RepetativeAttackDirector : AttackDirector
     // Public
 
     // Protected
+    protected Dictionary<ulong, SceneTreeTimer> enemyTimers = 
+		new Dictionary<ulong, SceneTreeTimer>(); 
 
     // Private
     private RepetativeAttackData data = null;
@@ -17,10 +20,23 @@ public partial class RepetativeAttackDirector : AttackDirector
     public override void _Ready()
     {
         base._Ready();
+
+        SetCollisionMaskValues();    
+
+		hitBoxDirector.AreaEntered += AddEnemyTimer;
+		hitBoxDirector.AreaExited += RemoveEnemyTimer;
+    }
+
+    public override void _Process(double delta)
+    {
+        base._Process(delta);
+
+        // Go through the dictionary and see which targets can take damage
+		CheckEnemyTimers();
     }
 
     //-------------------------------------------------------------------------
-	// Methods
+    // Methods
     // Public
     public override void LoadAttackDataFile() {
 		data = (RepetativeAttackData) GD.Load(dataPath);
@@ -31,6 +47,71 @@ public partial class RepetativeAttackDirector : AttackDirector
     }
 
     // Protected
+    protected void AddEnemyTimer(Area3D enemyArea) {
+		// Get the Opposing objects Interaction Director
+		InteractionDirector otherIteraction = 
+			enemyArea.GetNode<InteractionDirector>("..");
+
+		// Safety Check
+		if (otherIteraction == null) {
+			GD.Print($"{enemyArea.Name} had no Interaction Area");
+			return;
+		}
+		
+		// Take the Initial Damage
+		otherIteraction.TickHealth(GetAttackData().damage);
+
+		// Create the timer
+		SceneTreeTimer timer = GetTree().CreateTimer(GetAttackData().delay);
+		
+		// Log the Opposing Area and it's timer
+		enemyTimers.Add(otherIteraction.GetInstanceId(), timer);
+
+		if (debug) {
+			GD.Print($"{otherIteraction.GetInstanceId()} has entered the AoE");
+		}
+	}
+
+	protected void RemoveEnemyTimer(Area3D enemyArea) {
+		// Get the Opposing objects Interaction Director
+		InteractionDirector otherIteraction = 
+			enemyArea.GetNode<InteractionDirector>("..");
+
+		// Safety Check
+		if (otherIteraction == null) {
+			GD.Print($"{enemyArea.Name} had no Interaction Area");
+			return;
+		}
+
+		// Get the enemy's interaction dir Unique ID
+		ulong enemyID = otherIteraction.GetInstanceId();
+
+		// Remove the enemy's damage timer from the dictionary
+		enemyTimers.Remove(enemyID);
+
+		if (debug) {
+			GD.Print($"{enemyID} has left the AoE");
+		}
+	}
+
+    protected void CheckEnemyTimers() {
+		// Loop through the current list of enemy timers
+		foreach(KeyValuePair<ulong, SceneTreeTimer> enemyTimer in enemyTimers) {
+			if (enemyTimer.Value.TimeLeft <= 0) {
+				InteractionDirector enemyInteraction = 
+					(InteractionDirector) InstanceFromId(enemyTimer.Key);
+
+				if (enemyInteraction == null) {
+					enemyTimers.Remove(enemyTimer.Key);
+				}
+				
+				enemyInteraction.TickHealth(GetAttackData().damage);
+				
+				// Reset the EnemyTimer
+				enemyTimers[enemyTimer.Key] = GetTree().CreateTimer(GetAttackData().delay);
+			}
+		}
+	}
 
     // Private
